+++ spec/camlsql.sql_unknown_parameter_type.spec.js
@@ -1,0 +1,14 @@
+/**
+ * Tests to encode to internal field names
+ */
+var camlsql = require("../dist/public_html/js/camlsql.js");
+
+describe("Custom, unknown parameter Query #001", function() {
+
+ it("Custom parameter object 'Ost'", function() {
+     expect(function() {
+      camlsql.prepare("SELECT * FROM Notebook WHERE Author = ?", [{type : "Ost"}]).getXml();
+    }).toThrow("[camlsql] Parameter type is not not implemented Ost");
+  });
+
+});