+++ src/camlsql-js/sql-query/query-parser--list-and-fieldnames.js
@@ -35,8 +35,6 @@ function extractListAndFieldNameParts(workingObject) {
       }
 
       workingObject.query = m[3];
-    } else {
-      workingObject.query = "";
     }
   }
 }